refactor: extract ProxyHostDrawer tab components into separate files
Part of #103
